requirements.txt
tqdm
numpy
pandas
tensorflow-gpu
sklearn
keras-toolkit